requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
gunicorn==21.2.0
